requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.32.0
azure-mgmt-core==1.5.0
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.1.2
Django==4.2.18
django-celery-beat==2.8.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-rest-passwordreset==1.5.0
django-simple-history==3.8.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-writable-nested==0.7.2
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
flower==2.0.1
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isodate==0.7.2
kombu==5.5.4
model-bakery==1.20.4
msgpack==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==24.2
phonenumberslite==9.0.6
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode[pil]==7.4.2
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
toposort==1.10
tornado==6.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
gunicorn==21.2.0
